+++ src/blogpkg/incl/gitDt.nim
@@ -1,7 +1,7 @@
 from std/strformat import fmt
 from std/strutils import strip
 from std/osproc import execCmdEx
-from std/times import parse, format, DateTime
+from std/times import parse, format, DateTime, now
 
 const
   outTimeFormat = "yyyy/MM/dd hh:mm:ss tt"
@@ -16,8 +16,12 @@ proc prettyDt*(s: DateTime): string =
 
 proc gitFileLastModified*(f: string): DateTime =
   ## Gets the last modification date of a file in Git repository
-  fmt"git log -1 --format=%ci '{f}'".execCmdEx.output.parseGitDate
+  let res = fmt"git log -1 --format=%ci '{f}'".execCmdEx.output
+  result = if res.len > 0: parseGitDate res
+           else: now()
 
 proc gitFileCreation*(f: string): DateTime =
   ## Gets the creation date of a file in Git repository
-  fmt"git log --format=%ci '{f}' | tail -1".execCmdEx.output.parseGitDate
+  let res = fmt"git log --format=%ci '{f}' | tail -1".execCmdEx.output
+  result = if res.len > 0: parseGitDate res
+           else: now()